+++ test/test_exception_handling.py
@@ -1,9 +1,9 @@
-
-
-
+"""
+Module to provide tests for altering how the command line handles exceptions.
+"""
 import os
 from test.markdown_scanner import MarkdownScanner
-from test.utils import act_and_assert, create_temporary_configuration_file
+from test.utils import create_temporary_configuration_file
 
 
 def test_exception_handling_no_exception():
@@ -17,10 +17,12 @@ This triggers several rules:
 1. Bla
 """
 
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-            pass
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+        ) as file_name_3:
             # Arrange
             scanner = MarkdownScanner()
             supplied_arguments = ["scan", file_name_1, file_name_3]
@@ -32,7 +34,11 @@ This triggers several rules:
 {file_name_3}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 1; Actual: 0; Below] (blanks-around-headings,blanks-around-headers)
 {file_name_3}:1:2: MD010: Hard tabs [Column: 2] (no-hard-tabs)
 {file_name_3}:3:1: MD032: Lists should be surrounded by blank lines (blanks-around-lists)
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                "{file_name_1}", file_name_1
+            ).replace(
+                "{file_name_3}", file_name_3
+            )
             expected_error = ""
 
             # Act
@@ -42,6 +48,7 @@ This triggers several rules:
             execute_results.assert_results(
                 expected_output, expected_error, expected_return_code
             )
+
 
 def test_exception_handling_scan_with_plugin_exception_and_no_flag():
     """
@@ -61,24 +68,42 @@ throw_exception
         "test", "resources", "plugins", "bad", "bad_next_line_with_scan_trigger.py"
     )
 
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--add-plugin", plugin_path,"scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--add-plugin",
+                    plugin_path,
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """{file_name_1}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 1; Actual: 0; Below] (blanks-around-headings,blanks-around-headers)
 {file_name_1}:1:2: MD010: Hard tabs [Column: 2] (no-hard-tabs)
 {file_name_1}:3:1: MD032: Lists should be surrounded by blank lines (blanks-around-lists)
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """
                 
 BadPluginError encountered while scanning '{file_name_2}':
 (Line 3): Plugin id 'MDE008' had a critical failure during the 'next_line' action.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)
 
@@ -86,6 +111,7 @@ BadPluginError encountered while scanning '{file_name_2}':
                 execute_results.assert_results(
                     expected_output, expected_error, expected_return_code
                 )
+
 
 def test_exception_handling_scan_with_plugin_exception_and_flag():
     """
@@ -105,13 +131,26 @@ throw_exception
         "test", "resources", "plugins", "bad", "bad_next_line_with_scan_trigger.py"
     )
 
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--add-plugin", plugin_path,"--continue-on-error","scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--add-plugin",
+                    plugin_path,
+                    "--continue-on-error",
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """{file_name_1}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 1; Actual: 0; Below] (blanks-around-headings,blanks-around-headers)
@@ -120,9 +159,15 @@ throw_exception
 {file_name_3}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 1; Actual: 0; Below] (blanks-around-headings,blanks-around-headers)
 {file_name_3}:1:2: MD010: Hard tabs [Column: 2] (no-hard-tabs)
 {file_name_3}:3:1: MD032: Lists should be surrounded by blank lines (blanks-around-lists)
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """{file_name_2}:0:0: (Line 3): Plugin id 'MDE008' had a critical failure during the 'next_line' action.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)
 
@@ -130,6 +175,7 @@ throw_exception
                 execute_results.assert_results(
                     expected_output, expected_error, expected_return_code
                 )
+
 
 def test_exception_handling_scan_with_tokenization_exception_and_no_flag():
     """
@@ -146,23 +192,41 @@ This triggers several rules:
 test: assert
 ---
 """
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--set", extension_enable_front_matter, "scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--set",
+                    extension_enable_front_matter,
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """{file_name_1}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 1; Actual: 0; Below] (blanks-around-headings,blanks-around-headers)
 {file_name_1}:1:2: MD010: Hard tabs [Column: 2] (no-hard-tabs)
 {file_name_1}:3:1: MD032: Lists should be surrounded by blank lines (blanks-around-lists)
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """
                 
 Unexpected Error(BadTokenizationError): An unhandled error occurred processing the document.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)
 
@@ -170,6 +234,7 @@ Unexpected Error(BadTokenizationError): An unhandled error occurred processing t
                 execute_results.assert_results(
                     expected_output, expected_error, expected_return_code
                 )
+
 
 def test_exception_handling_scan_with_tokenization_exception_and_flag():
     """
@@ -186,13 +251,26 @@ This triggers several rules:
 test: assert
 ---
 """
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--continue-on-error","--set", extension_enable_front_matter, "scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--continue-on-error",
+                    "--set",
+                    extension_enable_front_matter,
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """{file_name_1}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 1; Actual: 0; Below] (blanks-around-headings,blanks-around-headers)
@@ -201,9 +279,15 @@ test: assert
 {file_name_3}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 1; Actual: 0; Below] (blanks-around-headings,blanks-around-headers)
 {file_name_3}:1:2: MD010: Hard tabs [Column: 2] (no-hard-tabs)
 {file_name_3}:3:1: MD032: Lists should be surrounded by blank lines (blanks-around-lists)
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """{file_name_2}:0:0: An unhandled error occurred processing the document.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)
 
@@ -211,6 +295,7 @@ test: assert
                 execute_results.assert_results(
                     expected_output, expected_error, expected_return_code
                 )
+
 
 def test_exception_handling_fix_with_plugin_exception_and_no_flag():
     """
@@ -230,22 +315,41 @@ throw_exception
         "test", "resources", "plugins", "bad", "bad_next_line_with_fix_trigger.py"
     )
 
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--add-plugin", plugin_path,"-x-fix","scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--add-plugin",
+                    plugin_path,
+                    "-x-fix",
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """Fixed: {file_name_1}
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """
                 
 BadPluginError encountered while scanning '{file_name_2}':
 (Line 3): Plugin id 'MDE008' had a critical failure during the 'next_line' action.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)
 
@@ -253,6 +357,7 @@ BadPluginError encountered while scanning '{file_name_2}':
                 execute_results.assert_results(
                     expected_output, expected_error, expected_return_code
                 )
+
 
 def test_exception_handling_fix_with_plugin_exception_and_flag():
     """
@@ -272,20 +377,40 @@ throw_exception
         "test", "resources", "plugins", "bad", "bad_next_line_with_fix_trigger.py"
     )
 
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--continue-on-error","--add-plugin", plugin_path,"-x-fix","scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--continue-on-error",
+                    "--add-plugin",
+                    plugin_path,
+                    "-x-fix",
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """Fixed: {file_name_1}
 Fixed: {file_name_3}
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """{file_name_2}:0:0: (Line 3): Plugin id 'MDE008' had a critical failure during the 'next_line' action.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)
 
@@ -293,6 +418,7 @@ Fixed: {file_name_3}
                 execute_results.assert_results(
                     expected_output, expected_error, expected_return_code
                 )
+
 
 def test_exception_handling_fix_with_tokenization_exception_and_no_flag():
     """
@@ -309,21 +435,40 @@ This triggers several rules:
 test: assert
 ---
 """
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--set", extension_enable_front_matter, "-x-fix","scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--set",
+                    extension_enable_front_matter,
+                    "-x-fix",
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """Fixed: {file_name_1}
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """
                 
 Unexpected Error(BadTokenizationError): An unhandled error occurred processing the document.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)
 
@@ -331,6 +476,7 @@ Unexpected Error(BadTokenizationError): An unhandled error occurred processing t
                 execute_results.assert_results(
                     expected_output, expected_error, expected_return_code
                 )
+
 
 def test_exception_handling_fix_with_tokenization_exception_and_flag():
     """
@@ -347,20 +493,40 @@ This triggers several rules:
 test: assert
 ---
 """
-    with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md") as file_name_1:
-        with create_temporary_configuration_file(contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md") as file_name_2:
-            with create_temporary_configuration_file(contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md") as file_name_3:
-
+    with create_temporary_configuration_file(
+        contents_file_1_and_3, file_name_prefix="tmp1", file_name_suffix=".md"
+    ) as file_name_1:
+        with create_temporary_configuration_file(
+            contents_file_2, file_name_prefix="tmp2", file_name_suffix=".md"
+        ) as file_name_2:
+            with create_temporary_configuration_file(
+                contents_file_1_and_3, file_name_prefix="tmp3", file_name_suffix=".md"
+            ) as file_name_3:
                 # Arrange
                 scanner = MarkdownScanner()
-                supplied_arguments = ["--continue-on-error","--set", extension_enable_front_matter, "-x-fix","scan", file_name_1, file_name_2, file_name_3]
+                supplied_arguments = [
+                    "--continue-on-error",
+                    "--set",
+                    extension_enable_front_matter,
+                    "-x-fix",
+                    "scan",
+                    file_name_1,
+                    file_name_2,
+                    file_name_3,
+                ]
 
                 expected_return_code = 1
                 expected_output = """Fixed: {file_name_1}
 Fixed: {file_name_3}
-""".replace("{file_name_1}", file_name_1).replace("{file_name_3}", file_name_3)
+""".replace(
+                    "{file_name_1}", file_name_1
+                ).replace(
+                    "{file_name_3}", file_name_3
+                )
                 expected_error = """{file_name_2}:0:0: An unhandled error occurred processing the document.
-""".replace("{file_name_2}", file_name_2)
+""".replace(
+                    "{file_name_2}", file_name_2
+                )
                 # Act
                 execute_results = scanner.invoke_main(arguments=supplied_arguments)