+++ changelog.md
@@ -45,6 +45,9 @@ the output of many of the existing rules, looking for missing issues.
   - Rule MD006 (disabled) - Added fix options
 - [Issue 941](https://github.com/jackdewinter/pymarkdown/issues/941)
   - Rule MD030 - Added fix options
+- [Issue 946](https://github.com/jackdewinter/pymarkdown/issues/946)
+  - Added `--continue-on-error` command line flag to "ignore" errors
+    and to keep on processing.
 
 ### Changed